requirements.txt
asgiref 
branca==0.4.2
certifi==2021.5.30
charset-normalizer==2.0.6
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django 
django-crispy-forms==1.13.0
django-import-export==2.6.1
et-xmlfile==1.1.0
folium==0.12.1
gunicorn==20.1.0
idna==3.2
Jinja2==3.0.2
MarkupPy==1.14
MarkupSafe==2.0.1
numpy==1.21.2
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.3
psycopg2 
python-dateutil==2.8.2
pytz 
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse 
tablib==3.0.0
typing-extensions 
urllib3==1.26.7
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
